api2/requirements.txt
fastapi
uvicorn
pyjwt
aiomysql
python-multipart
